Fix argument parsing for cargo execution
When invoking servum with cargo, as in `cargo servum example/` to
serve an example folder, the first user argument (after the
executable path) is `servum` and not `example/`.

If the program is executed using cargo, i.e. when installing with
cargo or using cargo run, the first argument will be skipped if it
is "servum". This means, that a folder `servum` cannot be served in
development mode using cargo run and without the ending slash,
which is bearable.