Fix argument parsing bug
This fixes a bug where servum is invoked using `cargo servum`. In
this case, the first argument, i.e. `servum` will be removed from
the argument vec. After this, the next if statement will try to
read from index 0 of an empty vec, result in a panic. Both branches
will now check, whether the vec is empty.